eval: metrics_impute

+++ KNNImputer/evaluation/metrics_impute.py
@@ -104,7 +104,7 @@ def elementwise(train_dataset, imputed):
     imputation = imputed.values[:, :C]
     imputation = imputation[train_dataset.mask[:, :C] == 1]
     
-    smape = np.abs(original - imputation)
+    smape = np.abs(original - imputation).astype(np.float32)
     smape /= (np.abs(original) + np.abs(imputation)) + 1e-6 # numerical stability
     smape = smape.mean()
